Added files via upload
Uploaded into this repo include the R code that was used along with the csv files that were created to be used to visualize the count of topics per candidate. Additionally the shinyT folder has the Shiny app that was created for Part two of the assignment to create a dynamic visualization.